+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## Changed
+
+- Get insertCoupon Mutation from checkout-resources
+
 ## [0.3.0] - 2019-09-04
 
 ## [0.2.2] - 2019-08-29