+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.6.0] - 2019-11-08
+
 ### Removed
 
 - `showPromoButton`, `setShowPromoButton`, `setCoupon` and `setCouponErrorKey` from the data provided by `OrderCouponProvider`.